+++ frontend/src/domains/components/FeedbackBoxBackGround/FeedbackBoxBackGround.tsx
@@ -1,7 +1,6 @@
 import { useAppTheme } from '@/hooks/useAppTheme';
-import { Theme } from '@/theme';
 import { FeedbackStatusType } from '@/types/feedbackStatus.types';
-import { css } from '@emotion/react';
+import { container } from './FeedbackBoxBackGround.styles';
 
 interface FeedbackBoxBackGroundProps {
   children: React.ReactNode;
@@ -16,16 +15,3 @@ export default function FeedbackBoxBackGround({
 
   return <section css={container(theme, type)}>{children}</section>;
 }
-
-export const container = (theme: Theme, type: FeedbackStatusType) => css`
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  width: 100%;
-  min-height: 100px;
-  padding: 18px;
-  border-radius: 14px;
-  box-shadow: 0 1px 3px 0 rgb(0 0 0 / 10%);
-
-  ${type === 'complete' && `background-color: ${theme.colors.gray[100]};`}
-`;
